Script_YT_Videos/YouTubeUploader/requirements.txt
customtkinter
patool
selenium
webdriver-manager
